Move AtWarning and AtError to their own file to avoid circular imports
