+++ pyat/at/lattice/exceptions.py
@@ -1,0 +1,10 @@
+"""Custom Warning and Error subclasses unique for AT.
+"""
+
+
+class AtError(Exception):
+    pass
+
+
+class AtWarning(UserWarning):
+    pass
